fix: close every finding of the 2026-08-17 audit
Four review passes over the whole repository - core concurrency, I/O, the
C# responder, security - and every finding fixed or priced as a documented
residual. The two criticals: the MCP SDK runs sync tools in worker threads
under a concurrent dispatcher, so the tool surface now serialises on an
explicit lock instead of a false premise; and the mutation area cap
computed its product in 32 bits, so settile:0,0,65536,65536,0 wrapped to
zero and rewrote the whole world - both cap sites now multiply in 64 bits.

Highs: replies carry a request id echoed by the responder (gated on a
published capability, so older vendored copies keep working untagged);
dispatch guards a throwing handler instead of crashing the game after the
trigger is consumed; settile checks TileLoader.TileCount like spawn and
give always did.

The rest, by module: log reads consume whole lines only and detect
rotation by fingerprint; snapshots survive a failed swap, refuse a failed
undo, and remove files a restore's snapshot never held; stop pins kills to
process creation time and refuses to fake success over a broken query;
stale-lock breaking and trigger release verify what they take before
taking it; the pre-arm clear checks addressing; consumption is an
exclusive move; the API index cache lands atomically; prune tolerates the
other session; and the sdist stops shipping docs/superpowers, whose
usernames and machine paths are also scrubbed from the test fixtures that
ride along in it.

Python 545 tests, C# 329, ruff clean, vendor manifest and template
re-synced.

+++ responder/DevMutationArgs.cs
@@ -269,8 +269,13 @@ namespace TModLoaderMcp.DevBridge
 				return false;
 			}
 
-			if (width * height > MaxArea) {
-				problem = width + " by " + height + " is " + (width * height) +
+			// The product in 64 bits, because the cap is what makes these verbs
+			// safe and 32-bit arithmetic is how a cap lies: 65536 x 65536 is
+			// 2^32, which IS zero as an int, and every product past 2^31 goes
+			// negative - both sail under any limit. A request that wraps the
+			// multiplication is precisely the request the limit exists to stop.
+			if ((long)width * height > MaxArea) {
+				problem = width + " by " + height + " is " + ((long)width * height) +
 					" tiles, past the limit of " + MaxArea + " one query may scan";
 				return false;
 			}
@@ -322,8 +327,10 @@ namespace TModLoaderMcp.DevBridge
 				return false;
 			}
 
-			if (width * height > MaxArea) {
-				problem = width + " by " + height + " is " + (width * height) +
+			// (long), for the reason TryResolveArea states: a product that wraps
+			// 32 bits reads as small or negative and walks straight past the cap.
+			if ((long)width * height > MaxArea) {
+				problem = width + " by " + height + " is " + ((long)width * height) +
 					" tiles, past the limit of " + MaxArea + " one request may place";
 				return false;
 			}
